MED-40 query to verify email from login page
