+++ apps/medon-fe/src/app/app.tsx
@@ -8,6 +8,7 @@ function App() {
   return (
     <Routes>
       <Route index path="/" element={<Login />} />
+      <Route path="/login" element={<Login />} />
       <Route path="/register" element={<RegistrationPage />} />
       <Route path="/forget-password" element={<ForgetPassword />} />
       <Route path="/reset-password" element={<ResetPassword />} />
